Move create opportunity form to modules
