+++ src/apps/investments/client/opportunities/Details/state.js
@@ -9,8 +9,6 @@ export const TASK_SAVE_OPPORTUNITY_REQUIREMENTS =
   'TASK_SAVE_OPPORTUNITY_REQUIREMENTS'
 
 export const TASK_SAVE_OPPORTUNITY_STATUS = 'TASK_SAVE_OPPORTUNITY_STATUS'
-export const TASK_CREATE_INVESTMENT_OPPORTUNITY =
-  'TASK_CREATE_INVESTMENT_OPPORTUNITY'
 
 export const ID = 'opportunityDetails'
