+++ src/apps/investments/client/opportunities/Details/tasks.js
@@ -84,9 +84,3 @@ export function saveOpportunityStatus({ opportunityId, values }) {
     status: values.status,
   })
 }
-
-export function createOpportunity(name) {
-  return apiProxyAxios
-    .post('v4/large-capital-opportunity', name)
-    .then(({ data: { id } }) => id)
-}
