+++ src/client/modules/Investments/Opportunities/CollectionList/metadata.js
